MSVC: fix warnings about conversion between float and double

+++ src/Gui/NavigationAnimation.cpp
@@ -66,7 +66,7 @@ void FixedTimeAnimation::initialize()
     float angle;
     SbRotation(navigation->getCamera()->orientation.getValue().inverse() * targetOrientation).getValue(rotationAxisPost, angle);
     if (angle > M_PI) {
-        angle -= 2 * M_PI;
+        angle -= float(2 * M_PI);
     }
 
     // Convert post-multiplication axis to a pre-multiplication axis
